requirements.txt
Django==1.3.1
PyYAML==3.10
beautifulsoup4==4.1.3
distribute==0.6.27
dj-database-url==0.2.1
nltk==2.0.4
numpy==1.6.2
psycopg2==2.4.5
python-Levenshtein==0.10.2
wsgiref==0.1.2
